Adding tests for SemVer 1.0.0 compatibility.
At the moment, the only test that cannot be passed is that of
prereleases with a single identifier containing leading zeroes.
This is permitted by the v1.0.0 spec, but not by v2.0.0.

I'm skipping over this test for the time being as an acceptable
loss. In the future, it will probably behoove us to have multiple
parser versions.