(RK-64) Remove build metadata from the library
Before this commit, the library could use build metadata in the version
numbers. After this commit, the build identifiers are not permitted in
the version.

The reason for this is that using this gem with Puppet would allow
puppet module generate to create versions with build identifiers. Those
modules would then by backward incompatible for no other reason than the
build metadata. Adding that breaking feature should be a choice.

Note: The code and tests have been left in, but disabled and commented,
so that it can be added back in when Semantic Version v2.0.0 is adopted
by PMT, R10K, and Forge at the same time.