(CODEMGMT-725) Externalize error message strings
This commit adds the ruby i18n libraries and externalizes all
error messages that the gem throws. When adding or changing
translation, the command `bundle exec rake gettext:pot`
must be run in order to create 'locales/semantic_puppet.pot'

Previously the gemspec required 'semantic_puppet', but that file
had to be changed to include the gettext functionality. This caused
an error when running any bundle command. The solution is to move the
information that the gemspec needs (the version number) into a separate
file (gem_version.rb) and include only that file.