(PUP-6373) Allow build identifiers in Version
Prior to this commit, it was not allowed to add build identifiers to
a version. This commit ensures that they can be used and that they do
not take part in the version comparison logic unless versions are
compared for equality.

This commit also improves performance by making the `#<=>` and `#to_s`
methods more efficient and by changing the hash calculation so that it
doesn't generate a new stringified version just to compute the hash.