(PUP-6373) Adjust spec tests with respect to pre-release identifiers
Before this commit, the spec tests assumed that versions with prerelease
identifiers should be included in ranges where no prerelease had been
defined for a given major, minor, patch tuple.

This commit changes the tests to that prereleases are included only if
the range explicitly requests a prerelease for the same major, minor,
patch tuple. Other prereleases are not included.