Add ability to use VersionRange with relaxed semantics
This commit adds a flag to the SemanticPuppet::VersionRange to allow
its semantics for how pre-releases are handled to be more forgiving.
The behavior if the new flag `strict_semver` is set to `false`, is the
same as for the previous version of the gem (0.1.4).

Backported from a723e23bd238ba5bf69e1d3e959b6aa93ec18211 in puppetlabs/puppet

References:
* https://tickets.puppetlabs.com/browse/PUP-7571
* https://tickets.puppetlabs.com/browse/PUP-7115

Signed-off-by: Markus Frosch <markus@lazyfrosch.de>