(maint) Fix travis and update ruby versions
Some specs would fail with jruby due to `.priority` being called on the
Source doubles from `module_release.rb`. Initialize the doubles with
predefined values.

It seems that this was done for some of the doubles in the test, but not
all. Also, I'm not sure why MRI passes and JRuby does not, as they use
the same rspec versions.